pyproject.toml
[project]
name = "stm32-coder-ft"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.46",
    "pdfplumber>=0.11.9",
    "pymupdf-layout>=1.27.1",
    "pymupdf4llm>=0.3.4",
    "trafilatura>=2.0.0",
    "unsloth>=2026.2.1",
]
